refactor(modeling - order) : 주문 패키지 구조 변경
